==> step0.index.R <==
#安装或加载所需要的packages
library(Seurat)
library(ggplot2)
library(clustree)
#install.packages("clustree",destdir = "D:/download_packages/")
library(cowplot)
library(dplyr)
suppressMessages(library(dplyr))
#install.packages("R.utils",destdir = "D:/download_packages/")
